코드 유연성 및 유지보수성 개선
### 주요 변경 사항
- `DropdownList.cs`: 메서드 이름 변경 및 리스너 간소화
- `ComponentBehaviour.cs`: `ConnectThisObject` 메서드 추가 및 초기화 코드 제거
- `ContainerObject.cs`: `IOdccObject` 타입으로 변경 및 메서드 추가
- `DataObject.cs`: `IsData` 속성 추가
- `IOdccItem.cs`: 네임스페이스 추가 및 인터페이스 상속 변경
- `OCBehaviour.cs`: 초기화 코드 간소화 및 메서드 호출 변경
- `ObjectBehaviour.cs`: 초기화 코드 변경 및 메서드 추가
- `SingletonObjectBehaviour.cs`: 싱글톤 초기화 및 파괴 로직 비활성화
- `OdccContainerTree.EditorFinder.cs`: 변수 이름 변경
- `OdccManager.cs`: 네임스페이스 추가 및 초기화/삭제 로직 수정
- `OdccUniversalDelegatesGenerator.cs`, `QuerySystem.cs`: 상속 관련 코드 제거
- `OdccFiniteStateMachine.cs` 등: 네임스페이스 변경
- `Test` 관련 파일: 조건부 컴파일 지시문 변경
- `manifest.json`, `packages-lock.json`: 패키지 버전 업데이트
- `ProjectSettings.asset`: `Scripting Define Symbols` 추가
- `ProjectVersion.txt`: Unity 에디터 버전 업데이트
- `ContainerNode.cs`: 새로운 클래스 추가 및 관련 메타데이터 정의

+++ Assets/Scritps/ODCC/Scripts/QuerySystem/OdccQueryCollector.cs
@@ -82,8 +82,7 @@ namespace BC.ODCC
 		public void RemoveLifeItem(object lifeItem)
 		{
 			if(lifeItem == null || IsDontDestoryLifeItem) return;
-			lifeItems.Remove(lifeItem);
-			if(lifeItems.Count == 0)
+			if(lifeItems.Remove(lifeItem) && lifeItems.Count == 0)
 			{
 				DeleteQueryCollector(this);
 			}
@@ -169,12 +168,12 @@ namespace BC.ODCC
 			{
 				OdccQueryCollector newCollector = new OdccQueryCollector(querySystem);
 
-				var objectList = OdccForeach.objectBehaviourList;
+				var objectList = OdccForeach.objectUpdateList;
 				if(objectList != null)
 				{
 					foreach(var item in objectList)
 					{
-						newCollector.AddObject(item);
+						newCollector.AddObject(item as ObjectBehaviour);
 					}
 				}
 
@@ -420,11 +419,11 @@ namespace BC.ODCC
 		/// </summary>
 		/// <param name="item">추가할 ObjectBehaviour 객체</param>
 		/// <param name="passDoubleCheck">중복 확인 여부</param>
-		internal void AddObject(ObjectBehaviour item, bool passDoubleCheck = false)
+		internal bool AddObject(ObjectBehaviour item)
 		{
-			if(queryItems.Contains(item)) return;
+			if(queryItems.Contains(item)) return true;
 
-			if(passDoubleCheck || IsSatisfiesQuery(item))
+			if(IsSatisfiesQuery(item))
 			{
 				var list = queryItems.ToList();
 				list.Add(item);
@@ -444,16 +443,18 @@ namespace BC.ODCC
 				onShowQueryItems = new List<ObjectBehaviour>();
 				onShowQueryItems.AddRange(list);
 #endif
+				return true;
 			}
+			return false;
 		}
 
 		/// <summary>
 		/// ObjectBehaviour 객체를 제거하는 메서드입니다.
 		/// </summary>
 		/// <param name="item">제거할 ObjectBehaviour 객체</param>
-		internal void RemoveObject(ObjectBehaviour item)
+		internal bool RemoveObject(ObjectBehaviour item)
 		{
-			if(!queryItems.Contains(item)) return;
+			if(!queryItems.Contains(item)) return true;
 
 			var list = queryItems.ToList();
 			if(list.Remove(item))
@@ -474,7 +475,9 @@ namespace BC.ODCC
 				onShowQueryItems = new List<ObjectBehaviour>();
 				onShowQueryItems.AddRange(list);
 #endif
+				return true;
 			}
+			return false;
 		}
 		/// <summary>
 		/// 수집기의 타겟이 유효할수 있는지 검사하는 메서드입니다.
@@ -515,7 +518,7 @@ namespace BC.ODCC
 			if(!Application.isPlaying) return false;
 #endif
 			if(item == null) return false;
-			if(item.ThisBehaviour.IsCallDestroy) return false;
+			if(item.ThisBehaviour.IsDestroy) return false;
 			if(querySystem == null) return false;
 			return querySystem.IsSatisfiesQuery(item);
 		}
@@ -526,14 +529,15 @@ namespace BC.ODCC
 		/// <param name="item">업데이트할 ObjectBehaviour 객체</param>
 		internal void UpdateObjectInQuery(ObjectBehaviour item)
 		{
-			if(IsSatisfiesQuery(item))
+			if(AddObject(item))
 			{
-				AddObject(item, true);
+				// Added
 			}
-			else
+			else if(RemoveObject(item))
 			{
-				RemoveObject(item);
+				// Removed
 			}
+			// Not Work
 		}
 
 		/// <summary>
